BookStore.UnitTests/Application/AuthorOperations/Commands/UpdateAuthor/UpdateAuthorCommandTest.cs
using AutoMapper;
using BookStore.Application.AuthorOperations.Commands.UpdateAuthor;
using BookStore.DBOperations;
using BookStore.UnitTests.TestSetup;
using FluentAssertions;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using Xunit;
using static BookStore.Application.AuthorOperations.Commands.UpdateAuthor.UpdateAuthorCommand;

namespace BookStore.UnitTests.Application.AuthorOperations.Commands.UpdateAuthor
{
    public class UpdateAuthorCommandTest : IClassFixture<CommonTestFixture>
    {
        private readonly BookStoreDbContext _context;
        private readonly IMapper _mapper;
        public UpdateAuthorCommandTest(CommonTestFixture testFixture)
        {
            _context = testFixture.Context;
            _mapper = testFixture.Mapper;
        }

        [Fact]
        public void WhenInvalidInputsAreGiven_InvalidOperationException_ShouldBeReturn()
        {
            //arrange
            UpdateAuthorCommand command = new UpdateAuthorCommand(_context, _mapper);
            UpdateAuthorModel model = new UpdateAuthorModel() {
                

            };
            command.Model = model;

            //act & assert
            FluentActions
                .Invoking(() => command.Handle(model))
                .Should().Throw<InvalidOperationException>().And.Message.Should().Be("Yazar bulunamadı!");
        }
        
            
        }
    }
